Beautiful Days at the Movies
Given a range of numbered days,[i...j]  and a number k , determine the number of days in the range that are beautiful. Beautiful numbers are defined as numbers where |i-reverse(i)| is evenly divisible by k . If a day's value is a beautiful number, it is a beautiful day. Print the number of beautiful days in the range.