Bon Appetit
Anna and Brian are sharing a meal at a restuarant and they agree to split the bill equally. Brian wants to order something that Anna is allergic to though, and they agree that Anna won't pay for that item. Brian gets the check and calculates Anna's portion. You must determine if his calculation is correct.
If Brian did not overcharge Anna, print Bon Appetit on a new line; otherwise, print the difference (i.e., b_charge-b_actual ) that Brian must refund to Anna. This will always be an integer.